requirements.txt
livekit-api
livekit-agents[google,silero,deepgram,turn-detector,elevenlabs]
python-dotenv
fastapi
uvicorn
langfuse<3.0.0
pillow
